+++ distributed/tests/test_worker.py
@@ -32,7 +32,6 @@ from distributed import (
     Client,
     Event,
     Nanny,
-    Reschedule,
     default_client,
     get_client,
     get_worker,
@@ -79,12 +78,9 @@ from distributed.worker import (
 from distributed.worker_state_machine import (
     AcquireReplicasEvent,
     ComputeTaskEvent,
-    Execute,
     ExecuteFailureEvent,
     ExecuteSuccessEvent,
-    FreeKeysEvent,
     RemoveReplicasEvent,
-    RescheduleEvent,
     SerializedTask,
     StealRequestEvent,
 )
@@ -1182,64 +1178,6 @@ async def test_get_current_task(c, s, a, b):
 
     result = await c.submit(some_name)
     assert result.startswith("some_name")
-
-
-@pytest.mark.slow
-@pytest.mark.parametrize("long_running", [False, True])
-@gen_cluster(
-    client=True,
-    nthreads=[("", 1)] * 2,
-    config={"distributed.scheduler.work-stealing": False},
-)
-async def test_reschedule(c, s, a, b, long_running):
-    a_address = a.address
-
-    def f(x):
-        if long_running:
-            distributed.secede()
-        sleep(0.1)
-        if get_worker().address == a_address:
-            raise Reschedule()
-
-    futures = c.map(f, range(4), key=["x1", "x2", "x3", "x4"])
-    futures2 = c.map(slowinc, range(10), delay=0.1, key="clog", workers=[a.address])
-    await wait(futures)
-    assert any(isinstance(ev, RescheduleEvent) for ev in a.state.stimulus_log)
-    assert all(f.key in b.data for f in futures)
-
-
-@gen_cluster(client=True, nthreads=[("", 1)])
-async def test_reschedule_released(c, s, a):
-    """A task raises Reschedule(), but the client previously released it"""
-    ev1 = Event()
-    ev2 = Event()
-
-    def f(ev1, ev2):
-        ev1.set()
-        ev2.wait()
-        raise Reschedule()
-
-    x = c.submit(f, ev1, ev2, key="x")
-    await ev1.wait()
-    x.release()
-    while "x" in s.tasks:
-        await asyncio.sleep(0.01)
-
-    await ev2.set()
-    while "x" in a.state.tasks:
-        await asyncio.sleep(0.01)
-
-
-def test_reschedule_released_worker_state(ws):
-    """Same as test_reschedule_released"""
-    instructions = ws.handle_stimulus(
-        ComputeTaskEvent.dummy(key="x", stimulus_id="s1"),
-        FreeKeysEvent(keys=["x"], stimulus_id="s2"),
-        RescheduleEvent(key="x", stimulus_id="s3"),
-    )
-    # There's no RescheduleMsg
-    assert instructions == [Execute(key="x", stimulus_id="s1")]
-    assert not ws.tasks
 
 
 @gen_cluster(nthreads=[])
